% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/approx.R
\name{approx_kinetics}
\alias{approx_kinetics}
\title{Approximates the kinetics of low order reactions with splines allowing the
mimicry of arbitrary non-linear kinetics.}
\usage{
approx_kinetics(reacts, concs, max_subs = 1, degree = 3,
  reversible = FALSE)
}
\arguments{
\item{reacts}{The reactions to be approximated}

\item{concs}{Vector of steady state concentrations}

\item{max_subs}{Only reactions with at most max_subs different substrates will be
approximated.}

\item{degree}{Polynomial degree of the approximation}

\item{reversible}{Whether reversible reactions are allowed or will be split.}
}
\value{
An R object conatining the following elements:
	\describe{
		\item{S}{The stochiometric matrix of the approximated system. Here
			the matrix will contain spoof reactions in order to realize the
			representation}
		\item{concs}{The approximated concentrations. For approximated reactions those
			are the values of the spline basis function at steady state concenyration.}
		\item{degree}{The polynomial degree of the approximation.}
		\item{approx}{Boolean vector indicating for each of the original reactions whether
				it has been approximated.}
}
}
\description{
Approximates the kinetics of low order reactions with splines allowing the
mimicry of arbitrary non-linear kinetics.
}

