% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{as.reactions}
\alias{as.reactions}
\title{Converts a stochiometric matrix to a reaction list.}
\usage{
as.reactions(s_matrix, reversible = F, r_names = NA)
}
\arguments{
\item{s_matrix}{The stochiometric matrix to be converted.}

\item{reversible}{Marks reversible reactions. FALSE denotes that
all reactions are irreversible. Otherwise a boolean vector of length
n_r defining the reversibility for each reaction.}
}
\value{
A reaction list containing the reactions from S.
}
\description{
Converts a stochiometric matrix to a reaction list.
}
\examples{
S = matrix(c(-1,1,1,-1))
rownames(S) <- c("A", "B")
print(as.reactions(S))
}
\keyword{reactions}
\keyword{stochiometry,}

