% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcone.R
\name{as.stability}
\alias{as.stability}
\title{Translates the eigenvectors to the corresponding stability term.}
\usage{
as.stability(evs)
}
\arguments{
\item{evs}{A vector of (complex) eigenvalues.}
}
\value{
A single string of either stable, unstable, limit cycle or 
 zero jacobian.
}
\description{
Translates the eigenvectors to the corresponding stability term.
}
\examples{
print(as.stability(c(0+2i, 0-3i)))
}
\seealso{
\code{\link{stability_analysis}} for an analysis of the entire
 basis.
}
\keyword{stability}

