% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linprog.R
\name{closest}
\alias{closest}
\title{Finds the closest point in within the k-cone to a query point p.}
\usage{
closest(p, S, m_terms)
}
\arguments{
\item{p}{The query point. Can be outside or inside the k-cone.}

\item{S}{The stochiometric matrix.}

\item{m_terms}{The metbaolic terms to be used. Must be have
\code{ncol(S)} elements.}
}
\value{
A vector containing the closest point to p within the k-cone.
}
\description{
Finds the closest point in within the k-cone to a query point p.
}
\examples{
data(eryth)
S <- stoichiometry(eryth)
mats <- runif(ncol(S))
closest(runif(ncol(S)), S, mats)

}
\seealso{
\code{\link{fba}} for optimization within the k-cone.
}

