% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{constant_flux}
\alias{constant_flux}
\title{Identifies reactions with a constant flux, thus reactions with no
substrates.}
\usage{
constant_flux(reacts)
}
\arguments{
\item{reacts}{The reaction list.}
}
\value{
A boolean vector identifying reactions with constant flux.
}
\description{
Identifies reactions with a constant flux, thus reactions with no
substrates.
}
\examples{
data(eryth)
which(constant_flux(eryth))
}
\seealso{
\code{\link{r_order}} to get the reaction order.
}
\keyword{reactions}

