% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcone.R
\name{constrain_by_Keq}
\alias{constrain_by_Keq}
\title{Get the equality constraints for a given set of known equlibrium constants.}
\usage{
constrain_by_Keq(r, name = "Keq", mats = NULL)
}
\arguments{
\item{r}{A reaction list.}

\item{name}{Name of the annotation that contains the constants.}

\item{mats}{A vector of mass-action terms to integrate with the equalities. Useful
if you want to constrain the flux cone rather than a k-cone. Default is
NULL (do not use mass-action terms).}
}
\value{
A n_r x n_e matrix containing the n_e constraints.
}
\description{
Get the equality constraints for a given set of known equlibrium constants.
}
\examples{
data(eryth)
x <- eryth
x[[4]]$Keq <- 10
co <- constrain_by_Keq(x)

}
\seealso{
\code{\link{polytope_basis}} how to add those to polytope
 computation.
}
\keyword{basis,}
\keyword{k-cone}

