% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{deriv_ma}
\alias{deriv_ma}
\title{Calculates derivatives of mass-action kinetics.}
\usage{
deriv_ma(i, substrates, concs)
}
\arguments{
\item{i}{The index of the substrate used as the differentiation variable.}

\item{substrates}{The stochiometry of the subbstrates (column in the
stochiometric matrix.}

\item{concs}{Concentrations of the substrates.}
}
\description{
Calculates derivatives of mass-action kinetics.
}
\examples{
data(eryth)
S = stochiometry(eryth)
dma1_dS3 = deriv_ma(3, S[,1], runif(nrow(S)))
}
\seealso{
\code{\link{jacobian}} to calculate the Jacobian matrix.
}

