% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcone.R
\name{enorm}
\alias{enorm}
\title{Calculates the euclidean norm of a vector}
\usage{
enorm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The norm of the vector x.
}
\description{
Calculates the euclidean norm of a vector
}
\examples{
x <- rnorm(10)
enorm(x)
}

