% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scraper.R
\name{find_hmdb}
\alias{find_hmdb}
\title{Finds IDs of compounds from HMDB corresponding to a given search term.}
\usage{
find_hmdb(search_term)
}
\arguments{
\item{search_term}{Can be any text search term or a different ID (for instance
a KEGG ID}
}
\value{
A single HMDB ID or list of putative IDs.
}
\description{
Finds IDs of compounds from HMDB corresponding to a given search term.
}

