% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{get_kcone_basis}
\alias{get_kcone_basis}
\title{Gets the null-space for the given k-cone assuming that all reactions
are irreversible.}
\usage{
get_kcone_basis(s_matrix, m_terms)
}
\arguments{
\item{s_matrix}{The stochiometric matrix to be used.}

\item{m_terms}{The metabolic terms.}
}
\value{
A matrix containing the basis vectors in the columns.
}
\description{
Gets the null-space for the given k-cone assuming that all reactions
are irreversible.
}
\examples{
data(eryth)
S = stochiometry(eryth)
K = kcone_null(S, runif(nrow(S)))
}
\seealso{
\code{\link{polytope_basis}} for arbitrary reversibility
 constraints.
}
\keyword{basis,}
\keyword{k-cone}

