% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{get_ma_terms}
\alias{get_ma_terms}
\title{Calculates all mass action terms \deqn{\prod_{i \in N^-} S_i^|N_i|}}
\usage{
get_ma_terms(s_matrix, concs)
}
\arguments{
\item{s_matrix}{The stochiometric matrix.}

\item{concs}{The concentrations of the substrates. Can be one of two
\itemize{
    \item A named vector containing the concentratiions for all substrates,
    where its names come from the same set as the rownames of S.
    \item A data frame with one column names "name" containing the
    names and all other columns containing multiple concentration entries.
}}
}
\value{
Either vector of length n_r containing the mass-action terms, if
	concs was a vector, or a matrix with n_r rows containing the mass-action
	in its columns.
}
\description{
Calculates all mass action terms \deqn{\prod_{i \in N^-} S_i^|N_i|}
}
\examples{
data(eryth)
S = stochiometry(eryth)
mats = ma_terms(S, runif(nrow(S)))
}
\seealso{
\code{\link{stochiometry()}} to calculate the stochiometric matrix.
}
\keyword{kinetics,}
\keyword{mass-action}

