% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{get_polytope_basis}
\alias{get_polytope_basis}
\title{Gets the polytope basis of the k-cone.}
\usage{
get_polytope_basis(s_matrix, v_terms = rep(1, ncol(s_matrix)))
}
\arguments{
\item{s_matrix}{The stochiometric matrix.}

\item{m_terms}{The metabolic terms for each reaction. Default to
calculation of flux cone.}
}
\value{
A matrix containing the basis vectors (normalized to a sum of 1)
	in its columns.
}
\description{
Gets the polytope basis of the k-cone.
}
\examples{
data(eryth)
V = polytope_basis(stochiometry(S))
}
\keyword{basis}
\keyword{k-cone,}

