% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{get_species}
\alias{get_species}
\title{Gets all species/metabolites from a reaction list.}
\usage{
get_species(reacts)
}
\arguments{
\item{react}{A reaction list.}
}
\value{
A vector containing all unique species in the model.
}
\description{
Gets all species/metabolites from a reaction list.
}
\examples{
data(eryth)
print(species(eryth))
}
\keyword{stochiometry}
\keyword{susbtrates,}

