% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{\%c\%}
\alias{\%c\%}
\title{%c% is a caching operator. It is used after an
expression, detects all assigned variables and saves
them into a file. If the file given exists the variables
are read from the cache file and not calculated again.
In order to delete the cache simply delete the associated
file.}
\usage{
ex \%c\% filename
}
\arguments{
\item{ex}{The R expression to be cached. Use brackets for
multiline statements.}

\item{filename}{The filename of the cache file.}
}
\value{
Nothing. Just caches all assigned variables in \code{filename}.
}
\description{
%c% is a caching operator. It is used after an
expression, detects all assigned variables and saves
them into a file. If the file given exists the variables
are read from the cache file and not calculated again.
In order to delete the cache simply delete the associated
file.
}
\examples{
# Is executed only if the file does not exist, otherwise
# samples a and b are read from 'samples.Rd'
{ a <- rnorm(1e5) } \%c\% 'samples.Rd'
}
\keyword{caching}
\keyword{operator,}
