% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{grep_id}
\alias{grep_id}
\title{Wrapper for grep that returns NA if nothing was found. Also splits up
comma sperated arguments.}
\usage{
grep_id(ids, x)
}
\arguments{
\item{ids}{ID or comma separated IDs to be searched.}

\item{x}{Character in which to search.}
}
\value{
The result of grep or NA if not found.
}
\description{
Wrapper for grep that returns NA if nothing was found. Also splits up
comma sperated arguments.
}
\examples{
grep_id('A, B', 'B, D')
}

