% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{grep_or_na}
\alias{grep_or_na}
\title{Wrapper for grep that returns NA if nothing was found.}
\usage{
grep_or_na(...)
}
\arguments{
\item{...}{Arguments passed on to grep}
}
\value{
The result of grep or NA if not found.
}
\description{
Wrapper for grep that returns NA if nothing was found.
}
\examples{
grep_or_na('A', 'blub')
}

