% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scraper.R
\name{hmdb_parse}
\alias{hmdb_parse}
\title{Parses basic data from a set of HMDB concentration nodes.}
\usage{
hmdb_parse(nodes)
}
\arguments{
\item{nodes}{A list of XML nodes.}
}
\value{
The parsed data set as a data frame.
}
\description{
Parses basic data from a set of HMDB concentration nodes.
}

