% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{hmdb_parse_xml}
\alias{hmdb_parse_xml}
\title{Parses the XML entry for a single HMDB id}
\usage{
hmdb_parse_xml(id)
}
\arguments{
\item{id}{The id for the entry}
}
\value{
The entry as a data frame
}
\description{
Parses the XML entry for a single HMDB id
}
