% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{jacobian}
\alias{jacobian}
\title{Calculates the Jacobian matrix.}
\usage{
jacobian(s_matrix, concs, deriv_func = deriv_ma)
}
\arguments{
\item{s_matrix}{The stochiometric matrix of the model.}

\item{concs}{The concentrations of the substrates.}

\item{deriv_func}{The function calculating the derivative. Must be of the
form func(idx, subs, concs) where idx is the index of the differentiation
variable, subs the stochiometry of the substrates and concs the concentration
of substrates. See \code{\link{mass_action}} for an example.}
}
\value{
The Jacobian matrix of dimension n_s x n_r where entry (i,j) denotes
	the derivative dv[j]/dS[i].
}
\description{
Calculates the Jacobian matrix.
}
\examples{
data(eryth)
S = stochiometry(eryth)
J = jacobian(S, runif(nrow(S)))
}
\seealso{
\code{\link{stochiometry}} to calculate the stochiometric matrix.
}
\keyword{jacobian,}
\keyword{stochiometry}

