% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{kcone}
\alias{kcone}
\title{Get the k-cone from a given flux cone.}
\usage{
kcone(V, mats, normalize = FALSE)
}
\arguments{
\item{V}{The flux cone basis. Rows denote the dimensions and columns the
individual basis vectors.}

\item{normalize}{Whether the basis vectors should be scaled to unit length.
Not recommened for differential analysis.}

\item{m_terms}{The metabolic terms. A vector with as many entries as rows in V.
All entries hsould be larger than 0.}
}
\value{
The stuff :O
}
\description{
\code{kcone()} builds the k-cone from a given flux cone basis V and a set
of metabolic terms M by calculating \eqn{M^{-1}V}.
}
\examples{
data(eryth)
S = stochiometry(eryth)
V = polytope_basis(S)
K = kcone(V, runif(ncol(S)))
}
\seealso{
\code{\link{polytope_basis}} for a way to calculate the flux cone,
 \code{\link{ma_terms}} to get mass action termsq
}
\keyword{basis}
\keyword{internal}

