% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{kcone_null}
\alias{kcone_null}
\title{Gets the null-space for the given k-cone assuming that all reactions
are irreversible.}
\usage{
kcone_null(s_matrix, m_terms)
}
\arguments{
\item{s_matrix}{The stochiometric matrix to be used.}

\item{m_terms}{The metabolic terms.}
}
\value{
A matrix containing the basis vectors in the columns.
}
\description{
Gets the null-space for the given k-cone assuming that all reactions
are irreversible.
}
\examples{
S <- matrix(c(1,0,0,1,-1,0, 0, -1), nrow=2)
rownames(S) = c('A', 'B')
K <- kcone_null(S, runif(ncol(S)))
}
\seealso{
\code{\link{polytope_basis}} for arbitrary reversibility
 constraints.
}
\keyword{basis,}
\keyword{k-cone}

