% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{make_irreversible}
\alias{make_irreversible}
\title{Converts a given reaction list into an irreversible one, splitting up
all reversible ones into two irreversible reactions.}
\usage{
make_irreversible(reacts)
}
\arguments{
\item{reacts}{The reaction list.}
}
\value{
A reaction list containing all reactions in reacts with reversible
 reactions being split in two.
}
\description{
Converts a given reaction list into an irreversible one, splitting up
all reversible ones into two irreversible reactions.
}
\examples{
data(eryth)
print(make_irreversible(eryth))
}
\keyword{reactions}

