% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{mass_action}
\alias{mass_action}
\title{Calculates the mass-action reaction rate}
\usage{
mass_action(substrates, concs)
}
\arguments{
\item{substrates}{Stochiometry of the substrates}

\item{concs}{The concentrations used for the substrates}
}
\value{
The mass action term \deqn{\prod_{i \in N^-} S_i^|N_i|}
}
\description{
Calculates the mass-action reaction rate
}
\examples{
data(eryth)
S = stochiometry(eryth)
ma1 = mass_action(S[,1], runif(nrow(S)))
}
\seealso{
\code{\link{stochiometry}} to generate the stochiometric
	matrix from a reaction list.
	\code{\link{ma_terms}} to calculate all mass-action terms at once.
}

