% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_perturbation.R
\name{minimal_perturbation}
\alias{minimal_perturbation}
\title{Assemble the LP problem for minimal perturbation.}
\usage{
minimal_perturbation(ma_terms, samples, reacts, permutations = 1000,
  v_min = 0, v_max = 1000, tradeoff = 0.95, min_obj = 100)
}
\arguments{
\item{ma_terms}{A matrix or data frame containing the metabolic terms in the
columns.}

\item{samples}{A factor or character string with either "normal" and
"disease" entries.}

\item{reacts}{The reactions for the respective model.}

\item{permutations}{The maximum number of permutations. If this number is
smaller than all possible permutation will check all permutations,
otherwise will sample that many permutations *with replacements*.}

\item{v_min}{The lower bounds for the fluxes. Either a single value for the
min bounds for all reactions or a matrix/data.frame with two columns and as
many rows as irreversible reactions where the columns contain the lower
bounds for each factor level in `samples`.}

\item{v_max}{The upper bounds for the fluxes. Either a single value for the
max bounds for all reactions or a matrix/data.frame with two columns and as
many rows as irreversible reactions where the columns contain the upper
bounds for each factor level in `samples`.}

\item{tradeoff}{Double in [0, 1] specifiying the equlibrium between
minimizing differences in k changes versus flux changes.}

\item{min_obj}{Constraints for growth rate. If a vector with two elements
the first will be interpreted as the index of the objective reaction and
the second value as its lower bound. If only a single value will constrain
the absolute sum of fluxes (sum |v_i|) and take the value as the lower
bound.}
}
\value{
A list with the following components:
 \describe{
 \item{fluxes}{The alterations in fluxes.}
 \item{k}{The alterations in kinetic constants kcat.}
 }
}
\description{
Assemble the LP problem for minimal perturbation.
}
\examples{
 NULL

}
