% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_perturbation.R
\name{minimal_perturbation_lp}
\alias{minimal_perturbation_lp}
\title{Assemble the LP problem for minimal perturbation.}
\usage{
minimal_perturbation_lp(ma_terms, samples, S, v_min = 0, v_max = 1000,
  tradeoff = 0.05, growth = NA)
}
\arguments{
\item{ma_terms}{A matrix or data frame containing the metabolic terms in the
columns.}

\item{samples}{A factor or character string with either "normal" and
"disease" entries.}

\item{S}{The stoichiometric matrix or a list containing one matrix for each
sample.}

\item{tradeoff}{Double in [0, 1] specifiying the equlibrium between
minimizing differences in k changes versus flux changes.}

\item{growth}{Constraints fro growth rate. If not NA must be a named vector
with entries "idx", "min", "max" denoting the index of the biomass reaction
and its minimum and maximum flux respectively.}

\item{bounds}{The bounds for the fluxes. Either a single value for the max
bound or a list each containing a vector specifying the bounds for the
specific sample.}
}
\value{
A list with the following components:
 \describe{
 \item{coefficients}{The coefficients for the equalities/inequalities in
     the LP. Each row denotes one equation.}
 \item{type}{The type of the (in)equalities. A character vector with as many
     entries as rows in `coefficients` being either "equal", "less" or
     "larger"}
 \item{bounds}{A matrix with two columns containing the upper and lower
     variable (flux) bounds.}
 \item{obj_coefs}{The objective coefficients. Containing one number for each
     variable.}
 }
}
\description{
Assemble the LP problem for minimal perturbation.
}
\examples{
 NULL

}
