% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{occupation}
\alias{occupation}
\title{Calculates the occupation for a basis. Here, occupation denotes
the faction of basis vectors for each varibale where the variable
is not zero.}
\usage{
occupation(basis)
}
\arguments{
\item{basis}{A basis matrix, where each column denotes a basis vector.}
}
\value{
A vector of length nrow(basis) containing the occupation for
	each dimension.
}
\description{
Calculates the occupation for a basis. Here, occupation denotes
the faction of basis vectors for each varibale where the variable
is not zero.
}
\examples{
S = matrix(c(1,0,0,1,-1,0, 0, -1), nrow=2)
rownames(S) = c("A", "B")
V = polytope_basis(S)
occ = occupation(V)
}
\seealso{
\code{\link{polytope_basis}} to get a k-cone.
}
\keyword{basis}

