% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scraper.R
\name{parse_conc}
\alias{parse_conc}
\title{Parses concentration values from HMDB.}
\usage{
parse_conc(val)
}
\arguments{
\item{val}{The value as it appears in HMDB.}
}
\value{
The numeric concentration value.
}
\description{
Parses concentration values from HMDB.
}

