% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linprog.R
\name{pfba}
\alias{pfba}
\title{Maximizes a set of given fluxes within the k-cone using parsimonous FBA.
Here, 'parsimonous' means that from all the possible maxima we will return
the one minimizing the overall flux sum, thus, being the most 'economic' one
for the organism.}
\usage{
pfba(obj, S, alpha = 1, v_min = 0, v_max = 1)
}
\arguments{
\item{obj}{The objective reaction, whose flux will be be maximized. Can
be any of the following three:
\itemize{
\item{A list containing at least the four vectors S, P, N_S, and N_P, which
    which contain the names of substrates and products and their respective
    stochiometries.}
\item{A unnamed vector containing the corresponding column in the
    stochiometric matrix.}
\item{A named vector containing only the non-zero entries in the respective
    column of the stochiometric matrix being named by their respective
    metabolite.}
\item{A single integer indicating which column of the stoichiometric
    contains the objective reaction.}
}}

\item{S}{The stochiometrix matrix to be used (must be irreversible).}

\item{alpha}{Fraction of optimum value. Objective must be at least
alpha * fba_solution.}

\item{v_min}{Lower bounds for the reaction fluxes. Can be a single value or a
vector containing one value for each reaction.}

\item{v_max}{Upper bounds for the reaction fluxes. Can be a single value or a
vector containing one value for each reaction.}
}
\value{
A vector with the same length as columns in \code{S} containing the
 solution of the optimization.
}
\description{
Maximizes a set of given fluxes within the k-cone using parsimonous FBA.
Here, 'parsimonous' means that from all the possible maxima we will return
the one minimizing the overall flux sum, thus, being the most 'economic' one
for the organism.
}
\examples{
S <- matrix(c(1, 0, -2, 1), ncol = 2)
rownames(S) <- c("A", "B")
pfba(c(B = -1), S)
}

