% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{plot.basis}
\alias{plot.basis}
\title{TODO: change me >:(}
\usage{
\method{plot}{basis}(b, n_cl = sqrt(ncol(b)/2), ...)
}
\arguments{
\item{...}{other arguments passed to specific methods}

\item{data}{default data set}
}
\value{
The stuff :O
}
\description{
\code{ggplot()} initializes a ggplot object. It can be used to
declare the input data frame for a graphic and to specify the
set of plot aesthetics intended to be common throughout all
subsequent layers unless specifically overridden.
}
\details{
There are three common ways to invoke \code{ggplot}:
\itemize{
   \item \code{ggplot(df, aes(x, y, <other aesthetics>))}
   \item \code{ggplot(df)}
   \item \code{ggplot()}
  }
}
\examples{
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)),
                 y = rnorm(30))
# Compute sample mean and standard deviation in each group
library(plyr)
ds <- ddply(df, .(gp), summarise, mean = mean(y), sd = sd(y))
}
\seealso{
\url{http://had.co.nz/ggplot2}
 \code{\link{geom_segment}} for a more general approach
}
\keyword{internal}

