% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{plot_basis}
\alias{plot_basis}
\title{Plots a heatmap of the basis.}
\usage{
plot_basis(b, n_cl = NA, ...)
}
\arguments{
\item{b}{The basis matrix, containing the basis vectors in the columns.}

\item{n_cl}{The number of clusters individual basis vectors are grouped
into. NA means no clustering is performed.}

\item{...}{other arguments passed to pheatmap.}
}
\description{
Plots a heatmap of the basis.
}
\examples{
S = matrix(c(1,0,0,1,-1,0, 0, -1), nrow=2)
rownames(S) = c("A", "B")
V = polytope_basis(S)
plot_basis(V)
}
\seealso{
\code{\link{plot_red}} for a geometric visualization.
}
\keyword{basis,}
\keyword{plot}

