% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{print.reactions}
\alias{print.reactions}
\title{Prints a nice description of the reactions.}
\usage{
\method{print}{reactions}(x)
}
\arguments{
\item{x}{The reaction list to be output.}
}
\description{
Prints a nice description of the reactions.
}
\examples{
data(eryth)
print(eryth)
}
\seealso{
\code{\link{as.reaction()}} to convert a stochiometric matrix
	to a reaction list.
}

