% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scraper.R
\name{priority_mean}
\alias{priority_mean}
\title{Extracts mean concentrations from HMDB results using a selection of biofluids
in a distinct order.}
\usage{
priority_mean(d, biofluids = c("cellular cytoplasm", "blood"))
}
\arguments{
\item{d}{A data frame as returned by hmdb_concentration}

\item{biofluids}{Vector of valid biofluids to be checked in the order
they appear.}
}
\value{
The mean
}
\description{
This is useful if one prefers concentrations from one biofluid, but will also
except other biofluids if no reported measurements can be found for the
primary one. \code{priority_mean} will return the mean of the first biofluid
for which measuremnts can be found.
}
\examples{
concs = hmdb_concentration("HMDB00124") # Fructose-6-phosphate
priority_mean(concs) # has measurements for cytoplasm so returns those
}
\seealso{
\code{\link{hmdb_concentration}} to parse the concentrations used
 in this function.
}

