% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{r_order}
\alias{r_order}
\title{Returns the number of different substrates a reaction has}
\usage{
r_order(reacts)
}
\arguments{
\item{reacts}{A reactions object as returned by \code{\link{read_reactions}}}
}
\value{
A numeric vector containing the number of substrates for each reaction
}
\description{
Returns the number of different substrates a reaction has
}

