% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{read_reactions}
\alias{read_reactions}
\title{Reads a list of reactions with optional annotations from a file.}
\usage{
read_reactions(react_file)
}
\arguments{
\item{react_file}{A csv file containing the reactions, where the first
   column denotes the reaction string and additional columns contain
annotations. A single annotation can contain several values as a
   comma separated string. The csv file should be saved with all entries
   quoted.}

\item{A}{list of reactions containing the stochiometry and additional
annotations.}
}
\description{
Reads a list of reactions with optional annotations from a file.
}
\examples{
r_str = "'reaction','name','numbers'\\n'A + B <=> C','r1','1,2,4'"
r = read_reactions(textConnection(r_str))
}

