% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sbml.R
\name{read_sbml}
\alias{read_sbml}
\title{Main driver to parse an SBML model.}
\usage{
read_sbml(sbml_file)
}
\arguments{
\item{sbml_file}{A SBML file. This can be a file on the disk or the location
to an online resource.}
}
\value{
A list containing three elements:
 \describe{
 \item{species}{The species (metabolites or molecules) in the model in a data
 frame. Boundary metabolites are removed automatically.}
 \item{params}{The global parameters defined for the model.}
 \item{reactions}{The reactions in the model as an "reactions" object.}
 }
}
\description{
This function wraps around \code{sbml_*} functions but also performs some
basic sanity checks and adjustments which are:
\itemize{
\item{Gives information about the used level and version and whether the FBC
 package is used.}
\item{Checks whether there are species that do not appear in reactions (warning).}
\item{Checks whether all species references in reactions are defined in the
 species list (error).}
\item{Maps lower and upper flux bounds to its respective global parameters.}
}
}
\examples{
# requires internet connection
m_url <- "http://www.ebi.ac.uk/biomodels-main/download?mid=BIOMD0000000042"
read_sbml(m_url)

}
\seealso{
\code{\link{sbml_species}}, \code{\link{sbml_params}} and
 \code{\link{sbml_reactions}} to read only parts of an SBML model.
}

