% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{rp}
\alias{rp}
\title{Obtains properties from a reaction list}
\usage{
rp(r, field = "KEGG_enzyme")
}
\arguments{
\item{r}{The reaction list.}

\item{field}{Name of the property to be obtained.}
}
\value{
A data.frame mapping the property to reaction indices.
}
\description{
Obtains properties from a reaction list
}

