% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scraper.R
\name{sbml_params}
\alias{sbml_params}
\title{Obtains the list of parameters from a SBML model}
\usage{
sbml_params(sbml_file)
}
\arguments{
\item{sbml_file}{A SBML file. This can be a file on the disk or the location
\tto an online resource.}
}
\value{
A data frame containing the obtained data. Some entries might be NA.
}
\description{
Obtains the list of parameters from a SBML model
}

