% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbml.R
\name{sbml_reactions}
\alias{sbml_reactions}
\title{Obtains the list of reactions from a SBML file.}
\usage{
sbml_reactions(sbml_file, progress = TRUE)
}
\arguments{
\item{sbml_file}{A SBML file. This can be a file on the disk or the location
to an online resource.}

\item{progress}{Whether a progress indication should be shown.}
}
\value{
A list with additional class "reactions" describing the reactions in
 the model. RDF annotations, notes and flux bounds are saved along with the 
 reactions
}
\description{
Obtains the list of reactions from a SBML file.
}
\examples{
# requires internet connection
m_url <- "http://www.ebi.ac.uk/compneur-srv/biomodels-main/download?mid=MODEL1103210001"
sbml_reactions(m_url)

}

