% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sbml.R
\name{sbml_species}
\alias{sbml_species}
\title{Obtains the list of species from a SBML model}
\usage{
sbml_species(sbml_file)
}
\arguments{
\item{sbml_file}{A SBML file. This can be a file on the disk or the location
to an online resource.}
}
\value{
A data frame containing the obtained data. Some entries might be NA.
}
\description{
Obtains the list of species from a SBML model
}
\examples{
# requires internet connection
m_url <- "http://www.ebi.ac.uk/compneur-srv/biomodels-main/download?mid=MODEL1103210001"
sbml_species(m_url)

}

