% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcone.R
\name{single_hyp}
\alias{single_hyp}
\title{Helper function to get the differential regulation between two approximations
of kinetic constansts.}
\usage{
single_hyp(k1, k2, reacts)
}
\arguments{
\item{k1}{A vector containing kinetic constants >=0.}

\item{k2}{A vector containing kinetic constants >=0.}

\item{reacts}{A reaction lists.}
}
\value{
A data frame containig the reaction indices, reactions, kind of regulation
 and log2-fold changes of k2 relative to k1. Log-fold changes where either of the
 constants is 0 are evaluated to 0.
}
\description{
Helper function to get the differential regulation between two approximations
of kinetic constansts.
}
\examples{
data(eryth)
n <- length(make_irreversible(eryth))
k1 <- runif(n)
k2 <- runif(n)
single_hyp(k1, k2, eryth)
}
\seealso{
\code{\link{hyp}} for a complete differential analysis.
}

