% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{species}
\alias{species}
\title{Gets all species/metabolites from a reaction list.}
\usage{
species(reacts)
}
\arguments{
\item{reacts}{A reaction list.}
}
\value{
A vector containing all unique species in the model.
}
\description{
Gets all species/metabolites from a reaction list.
}
\examples{
data(eryth)
print(species(eryth))
}
\keyword{stoichiometry}
\keyword{susbtrates,}

