% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcone.R
\name{stability_analysis}
\alias{stability_analysis}
\title{Performs a stability analysis for each basis vector in a k-cone.}
\usage{
stability_analysis(basis, s_matrix, concs)
}
\arguments{
\item{basis}{A basis matrix, where each column denotes a basis vector.}

\item{s_matrix}{The stochiometric matrix of the model.}

\item{concs}{A named vector of metabolite concentrations.}
}
\value{
A data frame containing the a string describing the type of
 stability along with its eigenvalues for each vector in the basis.
}
\description{
Performs a stability analysis for each basis vector in a k-cone.
}
\examples{
data(eryth)
S <- matrix(c(1,0,0,1,-1,0, 0, -1), nrow=2)
rownames(S) = c('A', 'B')
V <- polytope_basis(S)
concs <- runif(nrow(S))
names(concs) <- rownames(S)
K <- kcone(V, ma_terms(S, concs))
stab <- stability_analysis(K, S, concs)
}
\seealso{
\code{\link{polytope_basis}} and \code{}\link{kcone} to
 calculate the k-cone.
}
\keyword{stability}
