% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{stochiometry}
\alias{stochiometry}
\title{Calculates the stochiometric matrix for a list of reactions.}
\usage{
stochiometry(reacts, reversible = FALSE, const = NULL)
}
\arguments{
\item{reacts}{The reaction list to be used.}

\item{reversible}{Whether the stochiometric matrix can include reversible
\treactions.}

\item{const}{A vector of species names that are assumed to be invariant and
\twill be dropped from the stochiometric matrix. The default is not to drop
\tany species.}
}
\value{
The stochiometric matrix with dimension n_s x n_r.
}
\description{
Calculates the stochiometric matrix for a list of reactions.
}
\examples{
data(eryth)
S = stochiometry(eryth)
}
\seealso{
\code{\link{read_reactions}} to read a reaction list from a file.
}
\keyword{stochiometry}

