% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ode.R
\name{timecourse}
\alias{timecourse}
\title{Simulates the time course for a given model.}
\usage{
timecourse(x0, t, k, s_matrix)
}
\arguments{
\item{x0}{Initial concentrations of metabolites.}

\item{t}{A vector of monotonously increasing times where the simulation
is evaluated.}

\item{k}{The vector of kinetic constants used for simulation.}

\item{s_matrix}{The stochiometric matrix of the model.}
}
\value{
A matrix containing the time in the first column and temporal
 concentrations of the metabolites in the following columns.
}
\description{
Simulates the time course for a given model.
}
\keyword{ODE,}
\keyword{simulation}

