% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sbml.R
\name{write_sbml}
\alias{write_sbml}
\title{Writes a SBML model.}
\usage{
write_sbml(reacts, spec = NA, obj = NA, default_bounds = c(-1000, 1000),
  annmap = NA, out = "dycone_model.xml")
}
\arguments{
\item{reacts}{A reaction list.}

\item{spec}{Additional annotations for the species. Must be a data frame where
the first column denotes the name/id and additional columns are valid id
types from http://identifiers.org.}

\item{obj}{A positive integer denoting the index of the reaction taken to be
the optimization target. If NA no FBA objective will be added.}

\item{default_bounds}{numeric vector of length 2. The default bounds used if
no explicit lower and upper bounds are given for the reaction. If the reaction
is irreversible the lower bounds will be 0.}

\item{annmap}{An annotation map. A named character vector where the names denotes
reactions annotations in the reation list and the entries denote the respective
id on http://identifiers.org.}

\item{out}{A string denoting the name of the written SBML file.}
}
\value{
Nothing.
}
\description{
This function writes a reaction list as a SBML file. It accepts various levels
of additional information to help annotate the model. \code{write_sbml} will
only write SBML Level 3 Version 1 with FBC Version 2.
}
\examples{
data(eryth)
write_sbml(eryth, out="eryth.xml")

}
\seealso{
\code{\link{read_sbml}}, to read SBML models.
}

