% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linprog.R
\name{dba}
\alias{dba}
\title{Maximizes a weighted sum of fluxes within the k-cone.}
\usage{
dba(a, s_matrix, m_terms, lower = 0, upper = 1)
}
\arguments{
\item{a}{A vector a such that the fluxes \eqn{a^Tv} are maximized.}

\item{s_matrix}{The stochiometrix matrix to be used (must be irreversible).}

\item{m_terms}{The corresponding flux products.}

\item{lower}{Lower bound for the reaction constants.}

\item{upper}{Upper bound for the reactions constants.}
}
\value{
A vector with the same length as \code{a} containing the solution
 of the optimization.
}
\description{
Maximizes a weighted sum of fluxes within the k-cone.
}

