% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{deriv_ma}
\alias{deriv_ma}
\title{Calculates derivatives of mass-action kinetics.}
\usage{
deriv_ma(i, substrates, concs)
}
\arguments{
\item{i}{The index of the substrate used as the differentiation variable.}

\item{substrates}{The stoichiometry of the subbstrates (column in the
stochiometric matrix. Must be the same length as concs.}

\item{concs}{Concentrations of the substrates.}
}
\value{
Returns the derivative of the mass action kinetics for the reaction
 specified by \code{subtrates} differentiated by the \code{i}th variable.
}
\description{
\code{deriv_ma} does not check whether the
ordering of substrates in the stoichiometry and concs is corrects, so please
make sure that the indices in \code{substrates} and \code{concs} coincide.
}
\examples{
data(eryth)
S <- stoichiometry(eryth)
dma1_dS3 <- deriv_ma(3, S[,1], runif(nrow(S)))
}
\seealso{
\code{\link{jacobian}} to calculate the Jacobian matrix.
}

