% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eryth}
\alias{eryth}
\title{Metabolic network of human red blood cell.}
\format{A list of reactions.}
\source{
\url{http://journals.plos.org/plosone/article/asset?unique&id=
info:doi/10.1371/journal.pone.0004967.s003}
}
\usage{
eryth
}
\value{
Assigns the \code{eryth} reactions list to the namespace.
}
\description{
Contains the reactions of the human red blood cell (erythrocite) metabolic
model together with some randomly sampled rates k.
}
\keyword{datasets}

