% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linprog.R
\name{fva}
\alias{fva}
\title{Performs flux variance analysis for a given objective reaction. Thus, FVA
obtains lower and upper bounds for the fluxes under a given (sub-)optimal
solution.}
\usage{
fva(obj, alpha = 1, S, v_min = 0, v_max = 1)
}
\arguments{
\item{obj}{The objective reaction, whose flux will be be maximized. Can
be any of the following three:
\itemize{
\item{A list containing at least the four vectors S, P, N_S, and N_P, which
    which contain the names of substrates and products and their respective
    stochiometries.}
\item{A unnamed vector containing the corresponding column in the
    stochiometric matrix.}
\item{A named vector containing only the non-zero entries in the respective
    column of the stochiometric matrix bein named by their respective
    metabolite.}
}}

\item{alpha}{A positive scalar <=1. FVA is performed assuming that the
optimal solution can not be less than alpha*opt.}

\item{S}{The stochiometrix matrix to be used (must be irreversible).}

\item{v_min}{Lower bounds for the reaction fluxes. Can be a single value or a
vector containing one value for each reaction.}

\item{v_max}{Upper bounds for the reaction fluxes. Can be a single value or a
vector containing one value for each reaction.}
}
\value{
A data frame with two columns, min and max, denoting the respective
 minimum and maximum fluxes for each reaction.
}
\description{
Performs flux variance analysis for a given objective reaction. Thus, FVA
obtains lower and upper bounds for the fluxes under a given (sub-)optimal
solution.
}
\examples{
S <- matrix(c(1, 0, -2, 1), ncol = 2)
rownames(S) <- c("A", "B")
fva(c(B = -1), S=S)
}

