% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{hmdb_concentration}
\alias{hmdb_concentration}
\title{Scrapes measured metabolite concentration values for a given HMDB ID.}
\usage{
hmdb_concentration(hmids, add = NULL)
}
\arguments{
\item{hmids}{The HMDB ID or a vector of IDs for the metabolites.}

\item{add}{A data frame with as many rows as entries in hmid containing additional
information that will be added to the result (for instance names or ids).}
}
\value{
The scraped data set as a data frame or NULL if no concentrations were
 found.
}
\description{
Scrapes measured metabolite concentration values for a given HMDB ID.
}
\examples{
# requires internet connection
hmdb_concentration('HMDB00124')

}

