% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{hyp}
\alias{hyp}
\title{Identifies hypothesis for differentially regulated reactions between a set of
normal and disease conditions.}
\usage{
hyp(normal, disease, reacts, type = "transformation",
  correction_method = "fdr", sorted = T, v_opt = NULL, full = FALSE)
}
\arguments{
\item{normal}{A matrix or data frame containing the metabolic terms of the
normal condition in the columns.}

\item{disease}{A matrix or data frame containing the metabolic terms of the
diseasse condition in the columns.}

\item{type}{The type of analysis to be performed. Either 'transformation',
'optimization' or 'raw' for a pass-through option.}

\item{correction_method}{A correction method for the multiple test p-values.
Takes the same arguments as the method argument in p.adjust.}

\item{v_opt}{A vector defining the optimization criterion. Must have the same
length as there are irreversible reactions.}

\item{full}{If TRUE also returns the individual log fold changes along with
the differential regulation data.}

\item{sort}{Whether the results should be sorted by p-value and mean log-fold
change.}
}
\value{
If full is TRUE returns a list of generated hypothesis and the individual
	log fold changes between all reference basis and between reference and treatments.
	If full is FALSE only returns the generated hypothesis.
}
\description{
Identifies hypothesis for differentially regulated reactions between a set of
normal and disease conditions.
}

