% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{hyp}
\alias{hyp}
\title{Identifies hypothesis for differentially regulated reactions between a set of
normal and disease conditions.}
\usage{
hyp(normal, disease, reacts, correction_method = "fdr", full = FALSE)
}
\arguments{
\item{reacts}{A reaction list.}

\item{correction_method}{A correction method for the multiple test p-values.
Takes the same arguments as the method argument in p.adjust.}

\item{full}{If TRUE also returns the individual log fold changes along with
the differential regulation data.}

\item{ref_list}{A list of basis for the reference/control group.}

\item{treat_list}{A list of basis for the treatment/disease group.}
}
\value{
If full is TRUE returns a list of generated hypothesis and the individual
	log fold changes between all reference basis and between reference and treatments.
	If full is FALSE only returns the generated hypothesis.
}
\description{
Identifies hypothesis for differentially regulated reactions between a set of
normal and disease conditions.
}

