% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{hyp}
\alias{hyp}
\title{Identifies hypothesis for differentially regulated reactions between a set of
normal and disease conditions.}
\usage{
hyp(normal, disease, reacts, type = "transformation",
  correction_method = "fdr", sorted = T, v_opt = NULL, full = FALSE)
}
\arguments{
\item{normal}{A matrix or data frame containing the metabolic terms of the
normal condition in the columns.}

\item{disease}{A matrix or data frame containing the metabolic terms of the
diseasse condition in the columns.}

\item{reacts}{The reaction list.}

\item{type}{The type of analysis to be performed. Either 'transformation',
'optimization' or 'raw' for a pass-through option.}

\item{correction_method}{A correction method for the multiple test p-values.
\tTakes the same arguments as the method argument in p.adjust.}

\item{sorted}{Whether the results should be sorted by p-value and mean log-fold
change.}

\item{v_opt}{Only needs to be set if type=='optimization'. Defines a weights
specifying the optimization criterion. If F are the steady state fluxes of
the specified system, the optimization criterion will be maximization of
sum(v_opt*F). Must be the same length as there are irreversible reactions.}

\item{full}{If TRUE also returns the individual log fold changes along with
\tthe differential regulation data.}
}
\value{
If full is FALSE only returns the generated hypothesis as a data
 frame. If full is TRUE returns a list of generated hypothesis and the
 individual log fold changes between all reference basis and between
 reference and treatments. The full output will be a list with following
 elements:
\t\describe{
 \item{hyp}{The generated hypotheses together with statistics and reactions.}
 \item{lfc_normal}{The log2-fold changes of enzyme activity within the normal
     group for each of the irreversible reactions. Those are never sorted, so
     the first entry corresponds to the first reaction, etc.}
 \item{lfc_disease}{The log2-fold changes of enzyme activity within between
     the disease and normal group for each of the irreversible reactions.
     Also never sorted.}
 \item{obj_normal}{Only if type=='optimization'. The value of the objective
     function for each of the metabolic terms in 'normal'.}
 \item{obj_disease}{Only if type=='optimization'. The value of the objective
     function for each of the metabolic terms in 'disease'.}
 }
}
\description{
\code{hyp()} is the main driver for differential analysis. It first uses a
reduction method to approximate the fold-change of the kinetic constants. This
is followed by calculation of all combinatorial log-fold changes within the
normal group and between the disease and normal group. Finally, the wilcoxon
ranked sum statistics are obtained for each reaction and the p-values
corrected.
}
\keyword{analysis}
\keyword{hypothesis,}
\keyword{k-cone,}

