% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{mass_action}
\alias{mass_action}
\title{Calculates a mass-action reaction rate}
\usage{
mass_action(substrates, concs)
}
\arguments{
\item{substrates}{Stochiometry of the substrates. Must have the same length
\tas concs.}

\item{concs}{The concentrations used for the substrates}
}
\value{
The mass action term \deqn{\prod_{i \in N^-} S_i^|N_i|}
}
\description{
\code{mass_action} is mostly used inetrnally to get mass action rates for a
single column of the stochiometric matrix. It does not check whether the
ordering of substrates in the stochiometry and concs is corrects, so please
make sure that the indices in \code{substrates} and \code{concs} coincide.
}
\examples{
data(eryth)
S = stochiometry(eryth)
ma1 = mass_action(S[,1], runif(nrow(S)))
}
\seealso{
\code{\link{stochiometry}} to generate the stochiometric
\tmatrix from a reaction list. \code{\link{ma_terms}} for a wrapper that
\tcalculates all mass-action terms for a stochiometric matrix and also checks
\tfor correct ordering of the substrates.
\t\code{\link{ma_terms}} to calculate all mass-action terms at once.
}

