% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{mass_action}
\alias{mass_action}
\title{Calculates the mass-action reaction rate}
\usage{
mass_action(substrates, concs)
}
\arguments{
\item{substrates}{Stochiometry of the substrates}

\item{concs}{The concentrations used for the substrates}
}
\value{
The mass action term \deqn{\prod_{i \in N^-} S_i^|N_i|}
}
\description{
Calculates the mass-action reaction rate
}

