% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scraper.R
\name{patch}
\alias{patch}
\title{Patches holes in a data set by using group means and reference data.}
\usage{
patch(measurements, id, normal, treatment, ref_data = NULL)
}
\arguments{
\item{measurements}{A data frame with one id column, and at least one normal
and one treatment column}

\item{id}{An index or column name specifying the ids}

\item{normal}{A vector of indices or column names specifying the control
group}

\item{treatment}{A vector of indices or column names specifying the treatment
group}

\item{ref_data}{An optional data frame of reference data 2 or 3 columns,
where the first column denotes the ids, the second the normal reference data
and the (optional) third column the treatment reference data.}
}
\value{
The original measurement data frame with a maximum number of missing
data being patched.
}
\description{
Biological data is only rarely complete this, however cone analysis requires
fully populated steady state concentration data. This functions help to patch
holes in your data. It assumes that your measurement data set is data frame
where the id column uses a shared identifier as the reference data set and
the data columns specifify control and treatment measurements.
The optional reference data set is further used to fill up holes which cannot
b patched by group means.
}

