% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcone.R
\name{plot.basis}
\alias{plot.basis}
\title{Plots a heatmap of the basis.}
\usage{
\method{plot}{basis}(b, n_cl = NA, ...)
}
\arguments{
\item{b}{The basis matrix, containing the basis vectors in the columns.}

\item{n_cl}{The number of clusters individual basis vectors are grouped
into. NA means no clustering is performed.}

\item{...}{other arguments passed to pheatmap.}
}
\description{
Plots a heatmap of the basis.
}
\examples{
data(eryth)
V = polytope_basis(stochiometry(eryth))
plot(V, n_cl=sqrt(ncol(V)/2))
}
\seealso{
\code{\link{plot_red}} for a geometric visualization.
}
\keyword{basis,}
\keyword{plot}

