% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility.R
\name{plot.reactions}
\alias{plot.reactions}
\title{Plots the reactions as a graph.}
\usage{
\method{plot}{reactions}(x)
}
\arguments{
\item{x}{A reaction list.}
}
\description{
Plots the reactions as a graph.
}
\examples{
data(eryth)
plot(eryth)
}
\seealso{
\code{\link{print.reactions}} for a nicely formatted representation.
}
\keyword{plot,}
\keyword{reactions}

