% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility.R
\name{read_reactions}
\alias{read_reactions}
\title{Reads a list of reactions with optional annotations from a file.}
\usage{
read_reactions(react_file)
}
\arguments{
\item{react_file}{A csv file containing the reactions, where the first
column denotes the reaction string and additional columns contain
annotations. A single annotation can contain several values as a
comma separated string. The csv file should be saved with all entries
quoted.}
}
\value{
An object of with the S3 class \code{reactions} denoting a list
 of reactions.
}
\description{
Reads a list of reactions with optional annotations from a file.
}
\examples{
r_str <- "reaction,abbreviation,numbers\\nA -> B,blub,\\"1,2,3\\"\\nB <=>, bla, 3"
r <- read_reactions(textConnection(r_str))
}

