% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scraper.R
\name{sbml_species}
\alias{sbml_species}
\title{Obtains the list of species from a SBML model}
\usage{
sbml_species(sbml_file)
}
\arguments{
\item{sbml_file}{A SBML file. This can be a file on the disk or the location
to an online resource.}
}
\value{
A data frame containing the obtained data. Some entries might be NA.
}
\description{
Obtains the list of species from a SBML model
}
\examples{
# Not run because requires internet connection
# m_url <- "http://www.ebi.ac.uk/compneur-srv/biomodels-main/MODEL1103210001"
# sbml_species(m_url)
}

